pc-client/requirements.txt
PyQt5==5.15.10
requests==2.31.0
Pillow==10.0.1
